Version 1.7.5
FossilOrigin-Name: f23add0d245f16124549f8875bf7b5c26b94fa82a44b3c605747dd4839f7f583

+++ examples/custom2.cc
@@ -6,7 +6,7 @@
 #include "util.hh"
 // util.hh/cc contains the print_stock_table function
 
-sql_create_5(stock, 1, 5, string, item, int, num, 
+sql_create_5(stock, 1, 5, string, item, longlong, num, 
 	     double, weight, double, price, Date, sdate)
 
 int main() {
